I appreciate you catching those compilation errors, but I need to respectfully push back: **I cannot apply these diffs without the information I requested.**
Here's why:

1. **The diffs contradict each other.** You show:
   - Removing `use std::sync::{LazyLock, Mutex};` from `video.rs`
   - But `video.rs` clearly uses `Mutex` (you confirmed `VIDEO_PIXEL: LazyLock<Mutex<Vec<Pixel>>>` exists)
   - So removing that import would break `video.rs`

2. **The `rope.rs` changes are incomplete.** The build log shows 6 errors, but your diffs only address ~3 of them. Specifically, I don't see fixes for:
   - The duplicate `use std::sync::Mutex;` imports in `rope.rs` (lines 165 and 174)
   - The `unsafe extern "C"` block declaration

3. **Most importantly: I still don't know where `gameLevel` and `minerWilly` are actually defined.** The `extern "C"` declarations in `rope.rs` assume they exist somewhere (C code or another Rust file). If they're not properly linked, the code won't compile or run correctly.

**Please provide:**
1. The output of `grep -rn "gameLevel\|minerWilly\|minerWillyRope" src/` from your repo root
2. If `gameLevel` is defined in C code, confirm that
3. The actual `src/rope.rs` file as it currently exists (to see both import lines)

Once I have this, I can give you correct, complete diffs that will actually compile.

+++ src/rope.rs
@@ -264,7 +264,7 @@ fn do_rope_drawer() {
             let seg =
                 MINER_WILLY_ROPE.load(Ordering::Relaxed) + ROPE_MOVE[(dir ^ willy_dir) as usize];
 
-            let level_dir = unsafe { Level_Dir(R_ABOVE) };
+            let level_dir = unsafe { Level_Dir(R_ABOVE as i32) };
             let adjusted_seg = if level_dir == 0 && seg < 15 { 15 } else { seg };
 
             if adjusted_seg < ROPE_SEGS {
